// frontend/src/api/model/githubCom4H1RZooraInternalDomainImportJob.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */
import type { GithubCom4H1RZooraInternalDomainImportStatus } from './githubCom4H1RZooraInternalDomainImportStatus';
import type { GithubCom4H1RZooraInternalDomainImportType } from './githubCom4H1RZooraInternalDomainImportType';

export interface GithubCom4H1RZooraInternalDomainImportJob {
  created_at?: string;
  created_count?: number;
  error?: string;
  failed_count?: number;
  id?: string;
  media_id?: string;
  organization_id?: string;
  processed_rows?: number;
  skipped_count?: number;
  status?: GithubCom4H1RZooraInternalDomainImportStatus;
  total_rows?: number;
  type?: GithubCom4H1RZooraInternalDomainImportType;
  updated_at?: string;
  user_id?: string;
}